backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.3.9
autopep8==2.3.2
black==25.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cryptography==43.0.3
dill==0.3.9
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.9
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isodate==0.7.2
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.3.1
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.6
psycopg2==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.0
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.3.5
pyproj==3.7.1
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.25.1
six==1.16.0
sqlparse==0.5.2
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.9.0
